requirements.txt
RPi.GPIO==0.6.3
smbus==1.1
